Move DimmedLoadingAnimation to it's own file
